refactor: extract end-aligned scroll target helpers
